WindowLeak bugfix, onDestroy, LoAc Finish(MemoryUse) Merge branch 'master' of https://github.com/BecomeBeacon/BeaconLocker
# Conflicts:
#	app/src/main/java/com/example/becomebeacon/beaconlocker/MultiMapActivity.java

+++ app/src/main/java/com/example/becomebeacon/beaconlocker/BleDeviceInfo.java
@@ -319,7 +319,7 @@ public class BleDeviceInfo {
         this.nickname = nickname;
     }
 
-    public void setLimitDistance(int d)
+    public void setLimitDistance(double d)
     {
         this.limitDistance=d;
     }
